Fix the type of mk_wheel_interval
- xd75:markus has MK_W_INTERVAL_0 360, which exceeds the limit of uint8_t

+++ tmk_core/common/mousekey.c
@@ -58,7 +58,7 @@ uint8_t mk_time_to_max = MOUSEKEY_TIME_TO_MAX;
 //int8_t mk_curve = 0;
 /* wheel params */
 uint8_t mk_wheel_delay = MOUSEKEY_WHEEL_DELAY/10;
-uint8_t mk_wheel_interval = MOUSEKEY_WHEEL_INTERVAL;
+uint16_t mk_wheel_interval = MOUSEKEY_WHEEL_INTERVAL;
 uint8_t mk_wheel_max_speed = MOUSEKEY_WHEEL_MAX_SPEED;
 uint8_t mk_wheel_time_to_max = MOUSEKEY_WHEEL_TIME_TO_MAX;
